Dark1/requirements.txt
# Raspberry Pi 5 Multimodal LLM Assistant - Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
numpy>=1.24.0
Pillow>=10.0.0

# Hardware interfaces
spidev>=3.6
RPi.GPIO>=0.7.1
python3-rpi-lgpio>=0.2.0.0  # Alternative GPIO library
picamera2>=0.3.12

# LLM inference
llama-cpp-python>=0.2.0
# Note: For Raspberry Pi, you may need to build from source:
# CMAKE_ARGS="-DLLAMA_BLAS=ON -DLLAMA_BLAS_VENDOR=OpenBLAS" pip install llama-cpp-python

# Computer vision
ultralytics>=8.0.0  # YOLOv8
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.8.0

# Audio processing
pyaudio>=0.2.13
# Vosk (offline STT)
vosk>=0.3.45
# Alternative: whisper (more accurate, heavier)
# openai-whisper>=20231117
# Piper TTS (if available)
# piper-tts>=1.0.0
# Alternative: pyttsx3 (espeak wrapper)
pyttsx3>=2.90
# Porcupine wake word (optional, requires access key)
# pvporcupine>=3.0.0

# RAG and embeddings
sentence-transformers>=2.2.0
# Alternative vector store (optional)
# chromadb>=0.4.0
# faiss-cpu>=1.7.4

# Database
# SQLite is built into Python, no package needed

# Utilities
python-dateutil>=2.8.2

# Development (optional)
# pytest>=7.4.0
# black>=23.0.0

